Fix citation page duplicated entities
