+++ site/gatsby-site/page-creators/createEntitiesPages.js
@@ -1,6 +1,6 @@
-const { default: slugify } = require('slugify');
-
 const path = require('path');
+
+const { computeEntities } = require('../src/utils/entities');
 
 const createEntitiesPages = async (graphql, createPage) => {
   const {
@@ -17,78 +17,10 @@ const createEntitiesPages = async (graphql, createPage) => {
     }
   `);
 
-  const entititiesHash = {};
+  const entities = computeEntities({ incidents: incidents.nodes });
 
-  const fields = [
-    {
-      property: 'Alleged_deployer_of_AI_system',
-      key: 'incidentsAsDeployer',
-    },
-    {
-      property: 'Alleged_developer_of_AI_system',
-      key: 'incidentsAsDeveloper',
-    },
-  ];
-
-  for (const incident of incidents.nodes) {
-    for (const field of fields) {
-      for (const name of incident[field.property]) {
-        const id = slugify(name, { lower: true });
-
-        if (!entititiesHash[id]) {
-          entititiesHash[id] = {
-            id,
-            name,
-            incidentsAsDeveloper: [],
-            incidentsAsDeployer: [],
-            incidentsAsBoth: [],
-            relatedEntities: [],
-          };
-        }
-
-        if (!entititiesHash[id][field.key].includes(incident.incident_id)) {
-          entititiesHash[id][field.key].push(incident.incident_id);
-        }
-
-        if (fields.every((f) => entititiesHash[id][f.key].includes(incident.incident_id))) {
-          entititiesHash[id].incidentsAsBoth.push(incident.incident_id);
-        }
-      }
-    }
-  }
-
-  for (const id in entititiesHash) {
-    entititiesHash[id].relatedEntities = incidents.nodes
-      .filter((incident) =>
-        fields.some((field) =>
-          incident[field.property].map((p) => slugify(p, { lower: true })).includes(id)
-        )
-      )
-      .reduce((related, incident) => {
-        for (const field of fields) {
-          for (const name of incident[field.property]) {
-            const relatedId = slugify(name, { lower: true });
-
-            if (relatedId !== id && !related.some((r) => r.id == relatedId)) {
-              const { id, name, incidentsAsDeveloper, incidentsAsDeployer, incidentsAsBoth } =
-                entititiesHash[relatedId];
-
-              related.push({
-                id,
-                name,
-                incidents:
-                  incidentsAsDeveloper.length + incidentsAsDeployer.length - incidentsAsBoth.length,
-              });
-            }
-          }
-        }
-
-        return related;
-      }, []);
-  }
-
-  for (const id in entititiesHash) {
-    const entity = entititiesHash[id];
+  for (const entity of entities) {
+    const { id } = entity;
 
     const pagePath = `/entities/${id}`;
 
@@ -110,7 +42,7 @@ const createEntitiesPages = async (graphql, createPage) => {
     path: '/entities',
     component: path.resolve('./src/templates/entities.js'),
     context: {
-      entititiesHash,
+      entities,
     },
   });
 };
